resolved partially imported errors by extracting ntuples to utils folder.
