+++ rising/transforms/format.py
@@ -1,15 +1,9 @@
-import collections
-from itertools import repeat
 from typing import Callable, Dict, Hashable, Mapping, Sequence, Tuple, Union
 
-import numpy as np
-from torch import Tensor
+__all__ = ["MapToSeq", "SeqToMap", "PopKeys", "FilterKeys", "RenameKeys"]
 
+from rising.transforms.abstract import AbstractTransform
 from rising.transforms.functional.utility import filter_keys, pop_keys
-from .abstract import AbstractTransform
-
-__all__ = ["MapToSeq", "SeqToMap", "PopKeys", "FilterKeys", "RenameKeys", "ntuple", "single", "pair", "triple",
-           "quadruple"]
 
 
 class MapToSeq(AbstractTransform):
@@ -132,27 +126,3 @@ class RenameKeys(AbstractTransform):
         for old_key, new_key in self.keys.items():
             data[new_key] = data.pop(old_key)
         return data
-
-
-def ntuple(n):
-    def parse(x):
-        if isinstance(x, (Tensor, np.ndarray, str)):
-            return tuple(repeat(x, n))
-        if isinstance(x, collections.abc.Iterable):
-            item_list = list(x)
-            if len(item_list) == n:
-                return item_list
-            if len(item_list) == 1:
-                return tuple(repeat(item_list[0], n))
-            if len(list(x)) != n:
-                raise RuntimeError(f"Iterable shape inconsistent, n = {n}, given {len(list(x))}")
-
-        return tuple(repeat(x, n))
-
-    return parse
-
-
-single = ntuple(1)
-pair = ntuple(2)
-triple = ntuple(3)
-quadruple = ntuple(4)
